refactor(runtime): colocate types with producing operations
Move each measurement type to the module containing its producing
operation, with constructor visibility matching the trust boundary.

- Ticks, WallNs → time.rs (produced by read(), calibration)
- CpuNs → cpu_clock.rs (produced by cpu_now_ns())
- PollActive, PollDeltas → piano_future.rs (produced by begin_poll/end_poll)
- NameId stays in types.rs (boundary type, multiple entry points)

WallNs::from_raw and CpuNs::from_raw are now module-private.
Only the producing module can construct new values from raw integers.
Ticks::from_raw remains pub(crate) for inflight atomic deserialization.

Also: impl Add for WallNs replaces from_raw arithmetic in children.rs,
calibrate_bias() returns CpuNs directly instead of u64 bits.
